requirements.txt
llama-index-core
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-readers-file
chromadb
python-dotenv
pydantic-settings
sqlalchemy
fastapi
uvicorn[standard]
python-multipart
